[#19] refactor: error 체계 리팩토링
