+++ internal/room/adapter/in/http/error/room_error.go
@@ -4,8 +4,7 @@ import (
 	"errors"
 	"net/http"
 
-	"github.com/kjuiop/live-platform-go/internal/room/domain"
-	roomin "github.com/kjuiop/live-platform-go/internal/room/port/in"
+	roomerr "github.com/kjuiop/live-platform-go/internal/room/domain/errors"
 	serrors "github.com/kjuiop/live-platform-go/internal/shared/errors"
 )
 
@@ -26,15 +25,15 @@ type ErrMapping struct {
 
 func GetMapping(err error) ErrMapping {
 	switch {
-	case errors.Is(err, domain.ErrRoomNotFound):
-		return ErrMapping{http.StatusNotFound, ErrCodeRoomNotFound, "not found chat room"}
 	case errors.Is(err, serrors.ErrInvalidRequest):
 		return ErrMapping{http.StatusBadRequest, ErrCodeInvalidParams, "invalid request body"}
-	case errors.Is(err, roomin.ErrCreateChatRoomFailed):
+	case errors.Is(err, roomerr.ErrRoomNotFound):
+		return ErrMapping{http.StatusNotFound, ErrCodeRoomNotFound, "not found chat room"}
+	case errors.Is(err, roomerr.ErrCreateChatRoomFailed):
 		return ErrMapping{http.StatusInternalServerError, ErrCodeCreateFailed, "failed to create chat room"}
-	case errors.Is(err, roomin.ErrDeleteChatRoomFailed):
+	case errors.Is(err, roomerr.ErrDeleteChatRoomFailed):
 		return ErrMapping{http.StatusInternalServerError, ErrCodeDeleteFailed, "failed to delete chat room"}
-	case errors.Is(err, roomin.ErrGetChatRoomsFailed):
+	case errors.Is(err, roomerr.ErrGetChatRoomsFailed):
 		return ErrMapping{http.StatusInternalServerError, ErrCodeGetFailed, "failed to get chat rooms"}
 	default:
 		return ErrMapping{http.StatusInternalServerError, ErrCodeUnknown, "unknown error"}
