+++ internal/room/domain/room.go
@@ -1,7 +1,6 @@
 package domain
 
 import (
-	"errors"
 	"fmt"
 	"math/rand"
 	"strings"
@@ -9,10 +8,6 @@ import (
 
 	"github.com/kjuiop/live-platform-go/internal/room/adapter/in/http/form"
 	"github.com/kjuiop/live-platform-go/internal/shared/utils"
-)
-
-var (
-	ErrRoomNotFound = errors.New("room not found")
 )
 
 type RoomInfo struct {
